Refactored routes for nock testing and mobile testing
