+++ TripBorderFrontend/src/api/authAPI.js
@@ -1,5 +1,5 @@
 import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react';
-import { API_ROUTES } from '../constants/api';
+import { API_ROUTES } from '../constants/apiConstants';
 
 export const authAPI = createApi({
   reducerPath: 'authAPI',
