+++ TripBorderFrontend/src/api/usersAPI.js
@@ -1,5 +1,5 @@
 import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react';
-import { API_ROUTES } from '../constants/api';
+import { API_ROUTES } from '../constants/apiConstants';
 
 export const usersAPI = createApi({
   reducerPath: 'usersAPI',
